🚚 relocate card value view to card value store
